requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.9

openai==1.51.2

faster-whisper==1.0.3
ctranslate2==4.4.0

sentence-transformers==2.7.0
faiss-cpu==1.8.0

numpy==1.26.4

requests==2.32.3